again back to original path to datalab logo
the hub log showed that the hub webserver added a /hub/ prefix to
whatever path given and return a 404. The /hub/ prefix is not the
problem as this seems to be something internal mapped to the static hub
content in /usr/local/shared/static and it is part of the jupyter hub
image build. The problem seem to have fixed it self when I switched to
https.